requirements/base.txt
django==3.2
djangorestframework
django-environ
psycopg2-binary
django-stubs

